+++ energyquantified/data/periodseries.py
@@ -256,7 +256,7 @@ class Periodseries(Series):
         else:
             raise ValueError("Periodseries has no values")
 
-    def to_timeseries(self, frequency=None):
+    def to_timeseries(self, frequency=None, field="value"):
         """
         Convert this period-based series to a regular time series.
 
@@ -265,11 +265,18 @@ class Periodseries(Series):
 
         :param frequency: The frequency of the resulting time series
         :type frequency: Frequency, required
+        :param field: Which period attribute to use when converting to timeseries. Options:
+        'value' or 'installed'. Defaults to 'value'.
+        :type field: str, optional
         :return: A time series
         :rtype: Timeseries
         """
         # Verify parameters
         assert isinstance(frequency, Frequency), "Must be a frequency"
+        assert isinstance(field, str), "Must be a str"
+        assert field in tuple("value", "installed"), "field must be 'value' or 'installed'"
+        if field == "installed":
+            assert all(isinstance(p, CapacityPeriod) for p in self.data), "field='installed' requires a series of CapacityPeriods"
         # Prepare conversion
         resolution = Resolution(frequency, self.resolution.timezone)
         if not self.has_data():
@@ -285,7 +292,8 @@ class Periodseries(Series):
             periods=self.data,
             resolution=resolution,
             begin=begin,
-            end=end
+            end=end,
+            field=field,
         )
         # Convert
         data = [Value(dt, value) for dt, value in iterator]
@@ -298,7 +306,7 @@ class Periodseries(Series):
         timeseries.set_name(self._name)
         return timeseries
 
-    def to_df(self, frequency=None, name=None, single_level_header=False):
+    def to_df(self, frequency=None, name=None, single_level_header=False, field="value"):
         """
         Alias for :meth:`Periodseries.to_dataframe`.
 
@@ -319,6 +327,9 @@ class Periodseries(Series):
         :param single_level_header: Set to True to use single-level header \
             in the DataFrame, defaults to False
         :type single_level_header: boolean, optional
+        :param field: Which period attribute to use when converting to timeseries. Options:
+        'value' or 'installed'. Defaults to 'value'.
+        :type field: str, optional
         :return: A DataFrame
         :rtype: pandas.DataFrame
         :raises ImportError: When pandas is not installed on the system
@@ -326,11 +337,12 @@ class Periodseries(Series):
         return self.to_dataframe(
             frequency=frequency,
             name=name,
-            single_level_header=single_level_header
+            single_level_header=single_level_header,
+            field=field,
         )
 
     def to_dataframe(self, frequency=None, name=None,
-                     single_level_header=False):
+                     single_level_header=False, field="value"):
         """
         Convert this period-based to a ``pandas.DataFrame`` as a time series
         in the given frequency.
@@ -349,6 +361,9 @@ class Periodseries(Series):
         :param single_level_header: Set to True to use single-level header \
             in the DataFrame, defaults to False
         :type single_level_header: boolean, optional
+        :param field: Which period attribute to use when converting to timeseries. Options:
+        'value' or 'installed'. Defaults to 'value'.
+        :type field: str, optional
         :return: A DataFrame
         :rtype: pandas.DataFrame
         :raises ImportError: When pandas is not installed on the system
@@ -356,7 +371,7 @@ class Periodseries(Series):
         # Verify parameters
         assert isinstance(frequency, Frequency), "Must be a frequency"
         # Conversion
-        timeseries = self.to_timeseries(frequency=frequency)
+        timeseries = self.to_timeseries(frequency=frequency, field=field)
         df = timeseries.to_dataframe(
             name=name,
             single_level_header=single_level_header
@@ -384,11 +399,17 @@ class _PeriodsToTimeseriesIterator:
     A period-based series iterator used for conversions to Timeseries objects.
     """
 
-    def __init__(self, periods=None, resolution=None, begin=None, end=None):
+    def __init__(self, periods=None, resolution=None, begin=None, end=None, field="value"):
         self.periods = [p for p in periods if p.end > begin and p.begin < end]
         self.resolution = resolution
         self.begin = begin
         self.end = end
+        if field == "value":
+            self._get_value_func = lambda p: p.value
+        elif field == "installed":
+            self._get_value_func = lambda p: p.installed
+        else:
+            raise AssertionError("field must be 'value' or 'installed'")
         # Iterator stuff
         self.d = None
         self.p = None
@@ -419,7 +440,7 @@ class _PeriodsToTimeseriesIterator:
             return (d0, None)
         # Period covers the entire time interval
         if p.is_covered(d0, d1):
-            return (d0, p.value)
+            return (d0, self._get_value_func(p))
         # We do not have any values for given date
         if p.is_interval_before(d0, d1):
             return (d0, None)
@@ -469,7 +490,7 @@ class _PeriodsToTimeseriesIterator:
     def _calc_mean_periods(self, periods, begin, end):
         # Get value and duration in interval for each period
         available_weights = [
-            (p.value, p.get_duration_seconds(begin, end))
+            (self._get_value_func(p), p.get_duration_seconds(begin, end))
             for p in periods
         ]
         # Get sum of weight
@@ -500,7 +521,7 @@ class PeriodseriesList(list):
         # Asserts
         _validate_periodseries_list(iterable)
 
-    def to_timeseries(self, frequency=None):
+    def to_timeseries(self, frequency=None, field="value"):
         """
         Convert all period-based series in this list to time series.
 
@@ -509,6 +530,9 @@ class PeriodseriesList(list):
 
         :param frequency: The frequency of the resulting time series
         :type frequency: Frequency, required
+        :param field: Which period attribute to use when converting to timeseries. Options:
+        'value' or 'installed'. Defaults to 'value'.
+        :type field: str, optional
         :return: A list of time series
         :rtype: TimeseriesList
         """
@@ -516,11 +540,11 @@ class PeriodseriesList(list):
         assert isinstance(frequency, Frequency), "Must be a frequency"
         # Convert all period-based series to time series
         return TimeseriesList(
-            periodseries.to_timeseries(frequency=frequency)
+            periodseries.to_timeseries(frequency=frequency, field=field)
             for periodseries in self
         )
 
-    def to_df(self, frequency=None, single_level_header=False):
+    def to_df(self, frequency=None, single_level_header=False, field="value"):
         """
         Alias for :meth:`Timeseries.to_dataframe`.
 
@@ -533,16 +557,20 @@ class PeriodseriesList(list):
         :param single_level_header: Set to True to use single-level header \
             in the DataFrame, defaults to False
         :type single_level_header: boolean, optional
+        :param field: Which period attribute to use when converting to timeseries. Options:
+        'value' or 'installed'. Defaults to 'value'.
+        :type field: str, optional
         :return: A DataFrame
         :rtype: pandas.DataFrame
         :raises ImportError: When pandas is not installed on the system
         """
         return self.to_dataframe(
             frequency=frequency,
-            single_level_header=single_level_header
+            single_level_header=single_level_header,
+            field=field,
         )
 
-    def to_dataframe(self, frequency=None, single_level_header=False):
+    def to_dataframe(self, frequency=None, single_level_header=False, field="value"):
         """
         Convert this PeriodseriesList to a ``pandas.DataFrame`` where all time
         series are placed in its own column and are lined up with the date-time
@@ -553,6 +581,9 @@ class PeriodseriesList(list):
         :param single_level_header: Set to True to use single-level header \
             in the DataFrame, defaults to False
         :type single_level_header: boolean, optional
+        :param field: Which period attribute to use when converting to timeseries. Options:
+        'value' or 'installed'. Defaults to 'value'.
+        :type field: str, optional
         :return: A DataFrame
         :rtype: pandas.DataFrame
         :raises ImportError: When pandas is not installed on the system
@@ -560,7 +591,7 @@ class PeriodseriesList(list):
         # Verify parameters
         assert isinstance(frequency, Frequency), "Must be a frequency"
         # Convert to time series then to data frame
-        timeseries_list = self.to_timeseries(frequency=frequency)
+        timeseries_list = self.to_timeseries(frequency=frequency, field=field)
         return timeseries_list.to_dataframe(
             single_level_header=single_level_header
         )